Move CLI logic for rule systems to ExTTRPGDev.CLI.RuleSystems
