+++ lib/cli.ex
@@ -1,11 +1,7 @@
 # credo:disable-for-this-file Credo.Check.Warning.IoInspect
 defmodule ExTTRPGDev.CLI do
-  alias ExTTRPGDev.Characters.Character
-  alias ExTTRPGDev.RuleSystems
-  alias ExTTRPGDev.RuleSystems.Abilities
-  alias ExTTRPGDev.RuleSystems.Languages
-  alias ExTTRPGDev.RuleSystems.Skills
   alias ExTTRPGDev.CLI.Roll
+  alias ExTTRPGDev.CLI.RuleSystems
 
   @moduledoc """
   The CLI for the project
@@ -23,101 +19,8 @@ defmodule ExTTRPGDev.CLI do
         parse_double_dash: true,
         subcommands:
           Roll.commands() ++
+            RuleSystems.commands() ++
             [
-              list_systems: [
-                name: "list-systems",
-                about: "List systems that are setup to be used with ExTTRPGDev"
-              ],
-              system: [
-                name: "system",
-                about: "Top level command fo systems",
-                subcommands: [
-                  gen: [
-                    name: "gen",
-                    about: "Used for generating things for the system",
-                    subcommands: [
-                      stat_block: [
-                        name: "stat-block",
-                        about: "Generate stat blocks for characters of the system",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ],
-                      character: [
-                        name: "character",
-                        about: "Generate characters for system",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ]
-                    ]
-                  ],
-                  show: [
-                    name: "show",
-                    about: "Used for showing information about the rule system",
-                    subcommands: [
-                      abilities: [
-                        name: "abilities",
-                        about: "Show the rule systems character abilities",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ],
-                      languages: [
-                        name: "languages",
-                        about: "Show the rule systems languages",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ],
-                      metadata: [
-                        name: "metadata",
-                        about: "Show system metadata",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ],
-                      skills: [
-                        name: "skills",
-                        about: "Show rule system skills",
-                        args: [
-                          system: [
-                            value_name: "SYSTEM",
-                            help: "A supported system, e.g. dnd5e",
-                            required: true,
-                            parser: :string
-                          ]
-                        ]
-                      ]
-                    ]
-                  ]
-                ]
-              ],
               gen: [
                 name: "gen",
                 about: "system agnostic generation helpers",
@@ -139,11 +42,10 @@ defmodule ExTTRPGDev.CLI do
         Roll.handle(parse_result)
 
       {[:list_systems], _} ->
-        RuleSystems.list_systems()
-        |> IO.inspect(label: "Configured Systems")
+        RuleSystems.handle_list_systems()
 
       {[:system | sub_commands], parse_result} ->
-        handle_system_subcommands(sub_commands, parse_result)
+        RuleSystems.handle_system_subcommands(sub_commands, parse_result)
 
       {[:gen | sub_commands], _} ->
         handle_generate_subcommands(sub_commands)
@@ -159,87 +61,10 @@ defmodule ExTTRPGDev.CLI do
     end
   end
 
-  def handle_system_subcommands([command | subcommands], %Optimus.ParseResult{
-        args: %{system: system}
-      }) do
-    loaded_system =
-      system
-      |> RuleSystems.assert_configured!()
-      |> RuleSystems.load_system!()
-
-    case command do
-      :gen ->
-        handle_system_generation_subcommands(subcommands, loaded_system)
-
-      :show ->
-        handle_system_show_subcommands(subcommands, loaded_system)
-    end
-  end
-
-  def handle_system_generation_subcommands(
-        [command | _subcommands],
-        %RuleSystems.RuleSystem{} = system
-      ) do
-    case command do
-      :stat_block ->
-        RuleSystems.RuleSystem.gen_ability_scores_assigned(system)
-        |> IO.inspect()
-
-      :character ->
-        character = Character.gen_character!(system)
-        IO.puts("-- Name: #{character.name}")
-
-        Enum.each(character.ability_scores, fn {ability, scores} ->
-          IO.puts("#{ability}: #{Enum.sum(scores)}")
-        end)
-    end
-  end
-
-  def handle_system_show_subcommands(
-        [command | _subcommands],
-        %RuleSystems.RuleSystem{} = system
-      ) do
-    case command do
-      :abilities ->
-        show_abilities(system)
-
-      :languages ->
-        show_languages(system)
-
-      :metadata ->
-        Map.get(system, :metadata)
-        |> IO.inspect()
-
-      :skills ->
-        show_skills(system)
-    end
-  end
-
   def handle_generate_subcommands([command | _subcommands]) do
     case command do
       :name ->
         IO.inspect(Faker.Person.name())
     end
-  end
-
-  def show_abilities(%RuleSystems.RuleSystem{abilities: %Abilities{specs: specs}}) do
-    Enum.each(specs, fn %Abilities.Spec{name: name, abbreviation: abbr} ->
-      IO.puts("(#{abbr}) #{name}")
-    end)
-  end
-
-  def show_languages(%RuleSystems.RuleSystem{languages: languages}) do
-    Enum.each(languages, fn %Languages.Language{name: name, script: script} ->
-      IO.puts("Name: #{name}, Script: #{script}")
-    end)
-  end
-
-  def show_skills(%RuleSystems.RuleSystem{skills: skills} = system) do
-    Enum.each(skills, fn %Skills.Skill{name: name, modifying_stat: mod_stat} ->
-      %Abilities.Spec{abbreviation: abbr} =
-        RuleSystems.RuleSystem.get_spec_by_name(system, mod_stat)
-
-      IO.puts("(#{abbr}) #{name}")
-    end)
   end
 end
